refactor(payment): create a payment status function
