+++ src/components/Payment/PaymentHead.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export default styled.div`
+margin: 15px;
+width: 290px;
+font-size: 20px;
+line-height: 23px;
+color: #8e8e8e;
+`;
